pyproject.toml
[tool.poetry]
name = "nest"
version = "0.1.0"
description = "Automation and diagnostics tool for repair shops"
authors = ["Elite Repairs"]
readme = "README.md"
packages = [{ include = "nest" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
Pillow = "*"
psutil = "*"
beautifulsoup4 = "*"
tkcalendar = "*"
python-dateutil = "*"
WMI = "*"
pywin32 = "*"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
ruff = "^0.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]